ACS-6379: [HxI LI] S3 Credentials - PART II - renaming package storage.endpoint to storage.connector
